requirements.txt
numpy==1.14.5
https://download.pytorch.org/whl/cpu/torch-1.4.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.5.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
fastai==1.0.60
Pillow==7.1.1
streamlit==0.57.3
